Run tests as part of preview/updates
These tests depend on two things that aren't available in
PULUMI_TEST_MODE:

    1) Resource functions (like aws.getRegion()).

    2) Resource outputs (like the EKS cluster's version).

Instead of running without the Pulumi CLI in the picture --
which is necessary for (1) and (2) to work -- we can instead
run the tests from within the Pulumi program itself.

This has the extra benefit that tests will be run as part of
any update, and they are properly orchestrated with respect
to the engine's activities around previews and updates.

To run the tests, simply run

    $ pulumi preview

or just do an update as usual

    $ pulumi up

and you will see the test output emitted to the console.

We plan on reconciling this with PULUMI_TEST_MODE, and removing
the need to use the promise<T> function in here, as part of
https://github.com/pulumi/pulumi/issues/2287.